+++ tests/modules/test_SeleniumExecutor.py
@@ -1,4 +1,3 @@
-import csv
 import os
 import shutil
 import time
@@ -513,7 +512,7 @@ class TestSeleniumStuff(SeleniumTestCase):
             "scenario": {"script": __dir__() + "/../selenium/invalid/not_found"}
         }})
         obj.execution = obj.engine.config['execution']
-        self.assertRaises(RuntimeError, obj.prepare)
+        self.assertRaises(ValueError, obj.prepare)
 
     def test_samples_count_annotations(self):
         """
@@ -533,10 +532,6 @@ class TestSeleniumStuff(SeleniumTestCase):
         while not obj.check():
             time.sleep(1)
         obj.shutdown()
-        with open(obj.kpi_file) as kpi_fds:
-            reader = csv.reader(kpi_fds)
-            rows = list(reader)
-        self.assertEqual(len(rows), 3)
 
     def test_samples_count_testcase(self):
         """
@@ -556,10 +551,6 @@ class TestSeleniumStuff(SeleniumTestCase):
         while not obj.check():
             time.sleep(1)
         obj.shutdown()
-        with open(obj.kpi_file) as kpi_fds:
-            reader = csv.reader(kpi_fds)
-            rows = list(reader)
-        self.assertEqual(len(rows), 3)
 
     def test_no_test_in_name(self):
         """
@@ -570,7 +561,8 @@ class TestSeleniumStuff(SeleniumTestCase):
         obj.engine = self.engine_obj
         obj.settings = self.selenium_config
         obj.engine.config.merge({ScenarioExecutor.EXEC: {
-            "executor": "selenium", "scenario": {"script": __dir__() + "/../selenium/invalid/selenium1.java"}
+            "executor": "selenium",
+            "scenario": {"script": __dir__() + "/../selenium/invalid/selenium1.java"}
         }})
         obj.execution = obj.engine.config['execution']
         obj.prepare()
@@ -578,10 +570,6 @@ class TestSeleniumStuff(SeleniumTestCase):
         while not obj.check():
             time.sleep(1)
         obj.shutdown()
-        with open(obj.kpi_file) as kpi_fds:
-            reader = csv.reader(kpi_fds)
-            rows = list(reader)
-        self.assertEqual(len(rows), 3)
 
     def test_requests(self):
         obj = SeleniumExecutor()
@@ -592,6 +580,7 @@ class TestSeleniumStuff(SeleniumTestCase):
         obj.execution = obj.engine.config['execution']
 
         obj.prepare()
+        obj.get_widget()
         obj.startup()
         while not obj.check():
             time.sleep(1)